+++ src/etoolbox/utils/misc.py
@@ -1,5 +1,5 @@
 """Miscellaneous helpers and utilities."""
-from io import BytesIO
+from io import DEFAULT_BUFFER_SIZE, BytesIO
 from pathlib import Path
 
 import requests
@@ -36,4 +36,42 @@ def download(url: str, fname: Path | BytesIO) -> None:
     ):
         for data in resp.iter_content(chunk_size=1024):
             size = file.write(data)
+            bar.update(size)
+
+
+def ungzip(zip_path: Path, out_path=None):
+    """Un gzip a file with a progress bar.
+
+    Args:
+        zip_path: path of gzip archive.
+        out_path: path of resulting output file, if not provided will be the same as
+            zip_path without ``.gz`` suffix.
+
+    Returns: None
+
+    """
+    import gzip
+
+    tqdm = lazy_import("tqdm.auto", wait_for_signal=False)
+    tqdm_logging = lazy_import("tqdm.contrib.logging", wait_for_signal=False)
+
+    if out_path is None:
+        out_path = zip_path.parent / zip_path.stem
+
+    with (
+        tqdm_logging.logging_redirect_tqdm(),
+        gzip.open(zip_path, "rb") as zf,
+        open(out_path, "wb") as out,
+        tqdm.tqdm(
+            desc="Ungzipping " + zip_path.name,
+            unit="iB",
+            unit_scale=True,
+            unit_divisor=1024,
+        ) as bar,
+    ):
+        while True:
+            chunk = zf.read(DEFAULT_BUFFER_SIZE)
+            if not chunk:
+                break
+            size = out.write(chunk)
             bar.update(size)